Add Preventing Overwrite section to tutorial
Without this change, the tutorial code allows overwriting a square.
We add the code to prevent changing the value of a square if it has
already been set.
We also take the opportunity to refer the reader to the MDN glossary
article on truthiness.
We add a codepen for this new section.